update for conor
