+++ Posts/index.mjs
@@ -1,0 +1,4 @@
+export * from "/Posts/create";
+export * from "/Posts/read";
+export * from "/Posts/update";
+export * from "/Posts/delete";
