+++ authFetch.mjs
@@ -9,10 +9,9 @@ export function headers() {
     }
 }
 
-export async function authFetch(url, options) {
+export async function authFetch(url, options = {}) {
     return fetch(url, {
         ...options, 
         headers: headers()
     })
-
 }
