+++ post/index.html
@@ -6,7 +6,7 @@
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <title>Social Media PLatform Log in / registrer page </title>
+    <title>Social Media POST </title>
 
     <link rel="stylesheet" href="style2.css">
