--- posts.delete.mjs
@@ -1,1 +1,0 @@
-export async function removePost(id) {}
