--- posts.index.mjs
@@ -1,4 +1,0 @@
-export * from "/posts.create";
-export * from "/posts.read";
-export * from "/posts.update";
-export * from "/posts.delete";
