--- posts.read.mjs
@@ -1,4 +1,0 @@
-export async function getPosts() {}
-
-export async function getPost(id) {}
-
