+++ templates/post.mjs
@@ -1,0 +1,3 @@
+export function postTemplate(postData) {
+    return `div class="post" id=${}   `
+}
